refactor: replace custom card component
