+++ src/components/InventorySidebar.jsx
@@ -1,6 +1,6 @@
 import { memo, useMemo } from 'react'
 import PropTypes from 'prop-types'
-import Card from './Card'
+import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card'
 import { PencilIcon, TrashIcon } from '@heroicons/react/24/outline'
 
 function InventorySidebar({
@@ -25,38 +25,40 @@ function InventorySidebar({
   return (
     <nav className="flex flex-col space-y-2">
       {items.map((o) => (
-        <Card key={o.id} className="flex items-center justify-between p-2">
-          <div className="flex items-center flex-1">
-            <button
-              onClick={o.select}
-              className={`flex-1 text-left px-3 py-2 rounded hover:bg-primary/10 ${
-                selected?.id === o.id ? 'bg-primary/10 font-medium' : ''
-              }`}
-            >
-              {o.name}
-            </button>
-            {notifications[o.id] ? (
-              <span className="badge badge-error ml-2">
-                {notifications[o.id]}
-              </span>
-            ) : null}
-          </div>
-          <>
-            <button
-              onClick={o.edit}
-              className="ml-2 text-primary hover:text-primary/70"
-              title="Редактировать объект"
-            >
-              <PencilIcon className="w-4 h-4" />
-            </button>
-            <button
-              onClick={o.remove}
-              className="ml-2 text-red-500 hover:text-red-700"
-              title="Удалить объект"
-            >
-              <TrashIcon className="w-4 h-4" />
-            </button>
-          </>
+        <Card key={o.id}>
+          <CardHeader className="flex items-center justify-between p-2">
+            <div className="flex items-center flex-1">
+              <button
+                onClick={o.select}
+                className={`flex-1 text-left px-3 py-2 rounded hover:bg-primary/10 ${
+                  selected?.id === o.id ? 'bg-primary/10 font-medium' : ''
+                }`}
+              >
+                {o.name}
+              </button>
+              {notifications[o.id] ? (
+                <span className="badge badge-error ml-2">
+                  {notifications[o.id]}
+                </span>
+              ) : null}
+            </div>
+            <div className="flex items-center">
+              <button
+                onClick={o.edit}
+                className="ml-2 text-primary hover:text-primary/70"
+                title="Редактировать объект"
+              >
+                <PencilIcon className="w-4 h-4" />
+              </button>
+              <button
+                onClick={o.remove}
+                className="ml-2 text-red-500 hover:text-red-700"
+                title="Удалить объект"
+              >
+                <TrashIcon className="w-4 h-4" />
+              </button>
+            </div>
+          </CardHeader>
         </Card>
       ))}
     </nav>
